LG-4226: Add explicit flexible width button modifier (#197)
**Why**: As a consumer of the login.gov Design System, I expect to be able to override the default full-width button appearance on mobile to allow for flexible-width buttons, so that I can create interfaces where the button is shown adjacent to other text or controls.

Currently, all buttons are shown as full-width at mobile viewports. There are a few instances in the IDP of buttons appearing as "flexible width", with text or interactive content shown next to the button. These are not the majority, so an override option is included to let the developer choose to force the width to shrink as flexible, while retaining the current default behavior.

In discussing this with @anniehirshman-gsa , it was considered whether a similar exception override class should be included for minimum-width buttons (e.g. `usa-button--minimum-width`). There are currently no use-cases for this, so it was decided to defer this for future consideration if ever it becomes relevant.