Upgrade gulp-sass to v5, drop node-sass (#266)
* Upgrade gulp-sass to v5, drop node-sass

**Why**: Because node-sass has proven to be tricky to manage across different Node versions. We already _also_ depend on the sass package, which can be assigned as the compiler implementation in the latest version of gulp-sass. This also reduces the overall number of dependencies of the project.

Notably, the `sass` package is expected to be slower than `node-sass`, due to the fact that the latter compiles a C binary. In the future, we should look to migrate to the Dart version of the project, which is the currently maintained offering and offers similar performance to node-sass.

* Bump minimum Node engine to 12

Match gulp-sass

* Update test to import compiler direct from sass package

* Promisify common render function

clean-up, but fingers crossed maybe it could help build error? grasping at straws

* Synchronous sass render

Faster? According to docs: https://sass-lang.com/documentation/js-api