Update README.md for USWDS 3 usage guidance (#338)
Why? To tailor the document toward an audience which is primarily interested in using the design system on a Login.gov project, not contributing to the design system itself. This also favors documentation which describes a Sass compilation process, and new requirements of USWDS v3. It also recommends the identity-build-sass NPM package as an easier way to get started.