Improve styling for segmented "big" button group
Why? LGDS styles default button stroke width as 1px, and has custom styles to override stroke width to 2px for big buttons. USWDS button group styles use the stroke width, but are not aware of our customizations to double the width for big buttons, so we need to manually override these styles.